implementd a method for checking sunked ships.

+++ src/com/company/AddShips.java
@@ -1,13 +1,12 @@
 package com.company;
 
 public class AddShips {
-
     public static void inputText(int i) {
          String[] nameShip = {"Aircraft Carrier"};
          System.out.print("Enter the coordinates of the " + nameShip[0] + " (" + i + " cells):\n");
             }
 
-    public static void addShipsOnField(short startOfFirstCoordinates, short startOfTwoCoordinates, short endOfFirstCoordinates, short endOfTwoCoordinates, String[][] array) {
+    public static void addShipsOnField(int startOfFirstCoordinates, int startOfTwoCoordinates, int endOfFirstCoordinates,int endOfTwoCoordinates, String[][] array) {
         char[] al = Battleship.verticalMarkup();
         Battleship.HorizontalMarkup();
